Arreglo errores

+++ heap/pruebas_alumnos.c
@@ -123,7 +123,7 @@ void pruebas_heapsort() {
 }
 
 void pruebas_destruyendo(){
-    int largo = 10;
+    size_t largo = 10;
     heap_t* heap = heap_crear(comparar_void);
     int** array = malloc(sizeof(int*) * largo);
     for (int i = 0; i < largo; i++) {
